// my-app/src/Types.tsx
export interface inputData {
  id?: string;
  fname: string;
  mail: string;
  number: string;
  website: string;
  contactName: string;
  contactPhone: string;
  contactMail: string;
  notes: string;
  type: string;
  category: string;
  percentage: string;
  activeFrom: string;
  criticalAccount: string;
  paymentOptions: string;
}